+++ src/Exception/WebserviceRequestException.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Topdata\TopdataFoundationSW6\Exception;
+
+class WebserviceRequestException extends \RuntimeException
+{
+}