+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2026-07-10
+### Added
+- `WebserviceRequestException` and `WebserviceResponseException` exception classes
+- `CurlHttpClient` helper (moved from webservice-connector plugin)
+- `AbstractTopdataWebserviceClient` abstract base class for webservice access
+
 ## [1.2.3] - 2025-05-04
 ### Added
 - IntSet and StringSet helper classes